Added ability to view comments on a fish

+++ src/components/DeleteFishPrompt/DeleteFishPrompt.tsx
@@ -10,10 +10,16 @@ import {
   deleteDoc,
 } from 'firebase/firestore';
 import { db } from '../../Firebase';
-import './DeleteFishPrompt.css'
+import './DeleteFishPrompt.css';
 
 function DeleteFishPrompt(props: any) {
-  const {userObject, item, setDeletePrompt, fillProfileFishArray, setUserObjectFunc} = props;
+  const {
+    userObject,
+    item,
+    setDeletePrompt,
+    fillProfileFishArray,
+    setUserObjectFunc,
+  } = props;
 
   const handleDelete = async () => {
     const userDoc1 = await getDoc(doc(db, 'users', userObject.uid));
@@ -28,25 +34,57 @@ function DeleteFishPrompt(props: any) {
 
     const itemRef = doc(db, 'fish', item.fishID);
     const itemDoc = await getDoc(itemRef);
-    const itemObject = itemDoc.data();
-    await itemObject?.likes.forEach(async (element: any) => {
-      const q = query(
-        collection(db, 'users'),
-        where('username', '==', element),
-      );
-      const querySnapshot = await getDocs(q);
-      querySnapshot.forEach((userDoc: any) => {
-        const user = userDoc.data();
-        user.likes = user.likes.filter((ele: any) => ele !== item.fishID);
-        setDoc(doc(db, 'users', user.uid), user);
+    if (itemDoc.exists()) {
+      const itemObject = itemDoc.data();
+      await itemObject?.likes.forEach(async (element: any) => {
+        const q = query(
+          collection(db, 'users'),
+          where('username', '==', element),
+        );
+        const querySnapshot = await getDocs(q);
+        querySnapshot.forEach((userDoc: any) => {
+          const user = userDoc.data();
+          user.likes = user.likes.filter((ele: any) => ele !== item.fishID);
+          setDoc(doc(db, 'users', user.uid), user);
+        });
       });
-    });
-    deleteDoc(itemRef);
+      itemObject?.comments.forEach(async (element: any) => {
+        const q = query(collection(db, 'fish'), where('fishID', '==', element));
+        const querySnapshot = await getDocs(q);
+        querySnapshot.forEach((fishDoc: any) => {
+          deleteDoc(doc(db, 'fish', fishDoc.data().fishID));
+        });
+      });
+      if (itemObject.reply) {
+        const originalFishDoc = await getDoc(
+          doc(db, 'fish', itemObject.originalFishID),
+        );
+        if (originalFishDoc.exists()) {
+          const updatedOriginalFishObject = originalFishDoc.data();
+          updatedOriginalFishObject.comments =
+            updatedOriginalFishObject.comments.filter(
+              (element: string) => element !== itemObject.fishID,
+            );
+          setDoc(
+            doc(db, 'fish', itemObject.originalFishID),
+            updatedOriginalFishObject,
+          );
+        }
+      }
+      deleteDoc(itemRef);
+    }
+
     setUserObjectFunc();
   };
   return (
     <div className="absolute-background-div">
-      <div role="menu" tabIndex={0} className="delete-prompt-container" onKeyDown={(e) => e.preventDefault()} onClick={(e) => e.preventDefault()}>
+      <div
+        role="menu"
+        tabIndex={0}
+        className="delete-prompt-container"
+        onKeyDown={(e) => e.preventDefault()}
+        onClick={(e) => e.preventDefault()}
+      >
         <h3 style={{ color: 'black' }}>Delete Fish?</h3>
         <p>
           This can’t be undone and it will be removed from your profile, the
